feat: targeted E2E test resolver and runner
## E2E Test File Renames

Rename test files to match module kebab-case convention:
- player-page.spec.ts → player.spec.ts
- team-page.spec.ts → team.spec.ts
- team-stats.spec.ts → team-off-def-stats.spec.ts
- schedule.spec.ts → league-schedule.spec.ts
- api-v1.spec.ts → api.spec.ts
- login-logout.spec.ts → auth.spec.ts
- depth-chart.spec.ts → depth-chart-entry.spec.ts
- depth-chart-changes.spec.ts → depth-chart-entry-changes.spec.ts
- depth-chart-submission.spec.ts → depth-chart-entry-submission.spec.ts

## Leaderboards Split

Split leaderboards.spec.ts into:
- season-leaderboards.spec.ts (Season Leaderboards tests)
- career-leaderboards.spec.ts (Career Leaderboards tests)

## Convention-Based Resolver (bin/e2e-for-file)

- Extracts module name from PHP class path
- Converts PascalCase to kebab-case with acronym support
- Disambiguates prefix collisions (Player vs PlayerDatabase)
- Maps infrastructure modules to smoke tests

## Accumulated Test Runner (bin/e2e-changed)

- Reads deduplicated test paths from session temp file
- Supports --dry-run, --clear, --pid, and passthrough args
- Clears accumulator on successful run
